feat: export io character to its own widget
