+++ packages/io_crossword_ui/lib/src/widgets/widgets.dart
@@ -2,6 +2,7 @@ export 'card_scrollable_content_with_button.dart';
 export 'error_view.dart';
 export 'gemini_gradient.dart';
 export 'io_app_bar.dart';
+export 'io_character_field.dart';
 export 'io_crossword_card.dart';
 export 'io_layout.dart';
 export 'io_linear_progress_indicator.dart';
